requirements.txt
streamlit==1.38.0
pdfplumber==0.11.4
pandas==2.2.2
python-dateutil==2.9.0.post0
rapidfuzz==3.9.1
openpyxl==3.1.5
requests==2.32.3
